TESTS: start testing app dialogs (with 500ms timeout)
