+++ psychopy/tests/test_app/test_builder/test_Experiment.py
@@ -369,76 +369,3 @@ class TestExpt(object):
         assert namespace.makeLoopIndex('trials_2') == 'thisTrial_2'
         assert namespace.makeLoopIndex('stimuli') == 'thisStimulus'
 
-
-
-class Test_App(object):
-    """This test fetches all standard components and checks that, with default
-    settings, they can be added to a Routine and result in a script that compiles
-    """
-    @pytest.mark.usefixtures('pytest_namespace')
-    def setup(self):
-        self.app = pytest.app
-
-        self.builder = self.app.newBuilderFrame()
-        self.exp = self.builder.exp
-        self.here = path.abspath(path.dirname(__file__))
-        self.tmp_dir = mkdtemp(prefix='psychopy-tests-app')
-        self.exp.addRoutine('testRoutine')
-        self.testRoutine = self.exp.routines['testRoutine']
-        self.exp.flow.addRoutine(self.testRoutine, 0)
-
-    def teardown(self):
-        shutil.rmtree(self.tmp_dir, ignore_errors=True)
-
-    # def test_all_components(self):
-    #     for compName, compClass in list(allComponents.items()):
-    #         if compName in ['SettingsComponent']:
-    #             continue
-    #         thisComp = compClass(exp=self.exp, parentName='testRoutine', name=compName)
-    #         self._checkCompileWith(thisComp)
-
-    def test_BuilderFrame(self):
-        """Tests of the Builder frame. We can call dialog boxes using
-        a timeout (will simulate OK being pressed)
-        """
-        builderView = self.app.newBuilderFrame()
-
-        expfile = path.join(prefs.paths['tests'],
-                            'data', 'test001EntryImporting.psyexp')
-        builderView.fileOpen(filename=expfile)
-        builderView.setExperimentSettings(timeout=1000)
-        builderView.isModified = False
-        #assert exp == 5
-        del builderView
-
-
-#    def _send_OK_after(self):
-#        #this is supposed to help with sending button clicks but not working
-#        def clickOK():
-#            clickEvent = wx.CommandEvent(wx.wxEVT_COMMAND_BUTTON_CLICKED, wx.ID_OK)
-#            self.builder.ProcessEvent(clickEvent)
-#        wx.CallAfter(clickOK)
-
-    def _checkCompileWith(self, thisComp):
-        """Adds the component to the current Routine and makes sure it still
-        compiles
-        """
-        filename = thisComp.params['name'].val+'.py'
-        filepath = path.join(self.tmp_dir, filename)
-
-        self.testRoutine.addComponent(thisComp)
-        #make sure the mouse code compiles
-
-        # generate a script, similar to 'lastrun.py':
-        buff = self.exp.writeScript() # is a StringIO object
-        script = buff.getvalue()
-        assert len(script) > 1500 # default empty script is ~2200 chars
-
-        # save the script:
-        f = codecs.open(filepath, 'w', 'utf-8')
-        f.write(script)
-        f.close()
-
-        # compile the temp file to .pyc, catching error msgs (including no file at all):
-        py_compile.compile(filepath, doraise=True)
-        return filepath + 'c'
